pyproject.toml
[project]
name = "axolotl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "chat-exporter>=2.8.4",
    "discord-py>=2.6.4",
    "pyyaml>=6.0.3",
]
